+++ backends/rapidpro/msg.go
@@ -288,7 +288,7 @@ func downloadMediaToS3(ctx context.Context, b *backend, channel courier.Channel,
 		path = fmt.Sprintf("/%s", path)
 	}
 
-	s3URL, err := utils.PutS3File(b.s3Client, b.config.S3MediaBucket, path, mimeType, body)
+	s3URL, err := b.storage.Put(path, mimeType, body)
 	if err != nil {
 		return "", err
 	}